// src/types/entity/Product.ts
import BaseEntity from "./BaseEntity";
import Category from "./Category";

export default interface Product extends BaseEntity {
    name: string;
    unit: string;
    price: number;
    quantity: number;
    warrantyPeriod: number;
    isAvailable: boolean;
    photoURL: string;
    category?: Category;
}